#79: moved jasm! and AssemblyBlock
