+++ jodin-common/src/assembly/error.rs
@@ -1,6 +1,6 @@
-use crate::mvp::value::Value;
+use crate::assembly::value::Value;
 use std::error::Error;
-use std::num::{TryFromIntError};
+use std::num::TryFromIntError;
 use std::string::FromUtf8Error;
 use thiserror::Error;
