+++ jodin-common/src/core/mod.rs
@@ -12,5 +12,5 @@ pub mod tags;
 /// This should be used as the name of the callable native object that can be used within jodin programs.
 ///
 /// The native object is special because unlike other functions, the only type-checked parameter
-/// is the first, which must be a [`Value::String`](crate::mvp::value::Value) value.
+/// is the first, which must be a [`Value::String`](crate::assembly::value::Value) value.
 pub static NATIVE_OBJECT: &str = "__NATIVE";
