--- jodin-common/src/mvp.rs
@@ -1,6 +1,0 @@
-//! The minimum viable product for compilation
-
-pub mod bytecode;
-pub mod error;
-pub mod location;
-pub mod value;
