+++ jodin-rs-vm/src/loadables.rs
@@ -4,7 +4,6 @@ use crate::error::VMError;
 use crate::{VMTryLoadable, VirtualMachine};
 use jodin_common::unit::CompilationObject;
 
-
 use std::path::{Path, PathBuf};
 
 /// A regular file
