+++ jodin-rs-vm/tests/dicts.rs
@@ -1,7 +1,7 @@
-use jodin_common::mvp::bytecode::{Asm, Assembly};
-use jodin_common::mvp::location::AsmLocation;
-use jodin_common::mvp::value::Value;
-use jodin_common::{init_logging};
+use jodin_common::assembly::instructions::{Asm, Assembly};
+use jodin_common::assembly::location::AsmLocation;
+use jodin_common::assembly::value::Value;
+use jodin_common::init_logging;
 use jodin_rs_vm::core_traits::VirtualMachine;
 use jodin_rs_vm::function_names::RECEIVE_MESSAGE;
 use jodin_rs_vm::mvp::{MinimumALU, MinimumMemory};
