+++ jodinc/src/compilation/incremental.rs
@@ -22,15 +22,11 @@
 //!
 //! [#74]: https://github.com/joshradin/jodin-rs/issues/74
 
-use crate::compilation::jodin_vm_compiler::{
-    JodinVMCompiler,
-};
+use crate::compilation::jodin_vm_compiler::JodinVMCompiler;
 
 use crate::passes::analysis::analyze_with_preload;
 use crate::{optimize, JodinError};
-use jodin_common::compilation::{
-    Compilable, Compiler,
-};
+use jodin_common::compilation::{Compilable, Compiler};
 use jodin_common::compilation_settings::CompilationSettings;
 use jodin_common::parsing::parse_program;
 use jodin_common::unit::{CompilationObject, Incremental, TranslationUnit};
