+++ jodinc/src/compilation/jodin_vm_compiler/function_compiler.rs
@@ -1,19 +1,17 @@
 //! The micro-compiler for functions
 
-use crate::compilation::jodin_vm_compiler::asm_block::{
-    rel_label, temp_label, AssemblyBlock, InsertAsm,
-};
 use crate::compilation::jodin_vm_compiler::statement_compiler::StatementCompiler;
 use crate::compilation::jodin_vm_compiler::{invalid_tree_type, VariableUseTracker};
 use crate::compilation::JodinVM;
 use crate::{JodinError, JodinNode, JodinResult};
+use jodin_common::assembly::asm_block::{rel_label, temp_label, AssemblyBlock, InsertAsm};
 use jodin_common::ast::JodinNodeType;
 use jodin_common::compilation::MicroCompiler;
 use jodin_common::core::tags::TagTools;
 
-use jodin_common::mvp::bytecode::{Asm};
+use jodin_common::assembly::instructions::Asm;
 
-use jodin_common::mvp::value::Value;
+use jodin_common::assembly::value::Value;
 use std::cell::RefCell;
 use std::rc::Rc;
