+++ jodinc/src/compilation/jodin_vm_compiler/statement_compiler.rs
@@ -1,16 +1,17 @@
 //! The statement compiler
 
-use crate::compilation::jodin_vm_compiler::asm_block::{rel_label, AssemblyBlock, InsertAsm};
 use crate::compilation::jodin_vm_compiler::expression_compiler::ExpressionCompiler;
-use crate::compilation::jodin_vm_compiler::{VariableUseTracker};
+use crate::compilation::jodin_vm_compiler::VariableUseTracker;
 use crate::compilation::JodinVM;
-use crate::{jasm, JodinError, JodinNode, JodinResult};
+use crate::{JodinError, JodinNode, JodinResult};
+use jodin_common::assembly::asm_block::{rel_label, AssemblyBlock, InsertAsm};
+use jodin_common::assembly::instructions::Asm;
+use jodin_common::assembly::value::Value;
 use jodin_common::ast::JodinNodeType;
 use jodin_common::compilation::MicroCompiler;
 use jodin_common::error::JodinErrorType;
-use jodin_common::mvp::bytecode::Asm;
-use jodin_common::mvp::value::Value;
 
+use jodin_common::jasm;
 use std::cell::RefCell;
 use std::rc::Rc;
