+++ jodinc/src/compilation/mod.rs
@@ -3,13 +3,7 @@
 //!
 //! [JodinNode]: crate::ast::JodinNode
 
-
-
 // pub use c_compiler::{C99Compiler, C99};
-
-
-
-
 
 #[cfg(feature = "c_compiler")]
 pub mod c_compiler;
